+++ src/gcode/machine/MachineLinearizer.h
@@ -25,13 +25,7 @@
 
 namespace GCode {
   class MachineLinearizer : public MachineAdapter {
-    double maxArcError;
-
   public:
-    MachineLinearizer(double maxArcError = 0.01) : maxArcError(maxArcError) {}
-
-    void setMaxArcError(double x) {maxArcError = x;}
-
     // From MachineInterface
     void arc(const cb::Vector3D &offset, const cb::Vector3D &target,
              double degrees, plane_t plane);