+++ CHANGELOG.md
@@ -12,6 +12,8 @@
  - Allow dot in GCode variable reference names.
  - Warn about rendering empty workpiece.
  - Move all axes on "G28/G30 <axes>" as specified in RS274/NGC section 3.5.8.
+ - Segments straddle arc in linearization.
+ - Control max-arc-error with GCode var.
 
 ## v1.2.0:
  - Improved Buildbotics CNC controller connectivity.